[builders] graph builder split in sub builders
