+++ src/builders/geometry-builder.ts
@@ -2,13 +2,15 @@ import {Node} from "../graph/node"
 import {Point} from "../geometry/point"
 import {CurveTypes, CurveForRender, CurveForTest} from "../geometry/curve"
 
+import {NodeGeometries} from "../builders/graphbuilders/node-builder"
+
 export class GeometryBuilder {
 	newCurve(type: CurveTypes, firstNode: Node, secondNode: Node){
 		const points = [
-				firstNode.geometries.POINT_CENTER,
-				firstNode.geometries.POINT_RIGHT_HANDLE,
-				secondNode.geometries.POINT_LEFT_HANDLE,
-				secondNode.geometries.POINT_CENTER
+			firstNode.geometries[NodeGeometries.POINT__CENTER],
+			firstNode.geometries[NodeGeometries.POINT__RIGHT_HANDLE],
+			secondNode.geometries[NodeGeometries.POINT__LEFT_HANDLE],
+			secondNode.geometries[NodeGeometries.POINT__CENTER],
 		] as const;
 		if (type == CurveTypes.FOR_RENDER){
 			//@ts-ignore
