+++ src/builders/graph-builder.ts
@@ -8,43 +8,18 @@ import {Point} from "../geometry/point"
 import {CurveTypes, CurveForTest, CurveForRender} from "../geometry/curve"
 
 import {GeometryBuilder} from "../builders/geometry-builder"
+import {NodeBuilder} from "../builders/graphbuilders/node-builder"
+import {EdgeBuilder} from "../builders/graphbuilders/edge-builder"
 
 export class GraphBuilder{
-	graph: Graph = Graph.getInstance();
-	geometryBuilder: GeometryBuilder = new GeometryBuilder();
+	nodeBuilder: NodeBuilder = new NodeBuilder();
+	edgeBuilder: EdgeBuilder = new EdgeBuilder();
 
 	newNode(name: string, x: number, y: number){
-		const gameObjects: Record<string, GameObjectOnGraph> = {};
-		const geometries: Record<string, GeometryOnGraph> = {
-			'POINT_CENTER': new Point(x, y),
-			'POINT_LEFT_HANDLE': new Point(x, y),
-			'POINT_RIGHT_HANDLE': new Point(x, y)
-		}
-		const node = new Node(name, gameObjects, geometries)
-		this.graph.addNode(node);
-		Object.entries(geometries).forEach(([key, geometry]) => geometry.graphParentElement = node);
-		Object.entries(gameObjects).forEach(([key, gameObject]) => gameObject.graphParentElement = node);
-		return node;
+		return this.nodeBuilder.build(name, x, y);
 	}
 
 	newEdge(firstNode: Node, secondNode: Node){
-		const gameObjects: Record<string, GameObjectOnGraph> = {};
-		const geometries: Record<string, GeometryOnGraph> = {
-			'RENDER_CURVE': this.geometryBuilder.newCurve(
-				CurveTypes.FOR_RENDER, 
-				firstNode,
-				secondNode
-			),
-			'TEST_CURVE': this.geometryBuilder.newCurve(
-				CurveTypes.FOR_TEST, 
-				firstNode,
-				secondNode
-			),
-		}
-		const edge = new Edge(firstNode, secondNode, gameObjects, geometries);
-		this.graph.addEdge(edge);
-		Object.entries(geometries).forEach(([key, geometry]) => geometry.graphParentElement = edge);
-		Object.entries(gameObjects).forEach(([key, gameObject]) => gameObject.graphParentElement = edge);
-		return edge;
+		return this.edgeBuilder.build(firstNode, secondNode);
 	}
 }
